Refactor DTO package structure: move AuthResponse, LoginRequest, and RegisterRequest to request package; remove unused UserSettingsResponses class
