+++ finance-tracker-api/src/main/java/com/tamthong/finance_tracker_api/dto/request/RegisterRequest.java
@@ -1,4 +1,4 @@
-package com.tamthong.finance_tracker_api.dto;
+package com.tamthong.finance_tracker_api.dto.request;
 
 import lombok.Data;
