+++ finance-tracker-api/src/main/java/com/tamthong/finance_tracker_api/service/AuthService.java
@@ -1,8 +1,9 @@
 package com.tamthong.finance_tracker_api.service;
 
-import com.tamthong.finance_tracker_api.dto.AuthResponse;
-import com.tamthong.finance_tracker_api.dto.LoginRequest;
-import com.tamthong.finance_tracker_api.dto.RegisterRequest;
+import com.tamthong.finance_tracker_api.dto.request.AuthResponse;
+import com.tamthong.finance_tracker_api.dto.request.LoginRequest;
+import com.tamthong.finance_tracker_api.dto.request.RegisterRequest;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
